feat(metadata):  singleton metadata can now indicate root container target
When there’s a hierarchy of child containers, normally registering a
singleton just registers it as singleton inside the child container
itself. However, you may want a singleton for the entire hierarchy.
This addition allows metadata to declare that the singleton should be
on the root. This enables the router and compose system to re-use
view-model instance across navigation/composition so that they
maintain state for the life of the application.

Fixes https://github.com/aurelia/router/issues/21