fix(singleton): alway register in the root container
This is a potential BREAKING CHANGE. The singleton metadata will now
always register the type in the root container by default. To override
this, provide a boolean value as the first or second argument of true
to register with the current child container only. Not that that would
be the default for autoregistration, so it’s not normally necessary to
state it explicitly.

Fixes #22